# api/openapi.yaml
openapi: 3.1.0
info:
  title: AgentFlow API
  version: 0.1.0
  description: |
    Authoritative HTTP contract shared by generated Go DTOs and the typed
    TypeScript client. Business handlers remain implemented with net/http.
    With AUTH_MODE=oidc, every business operation requires a revocable session
    cookie and owned/granted Workspace entity (401/404 on failure). Mutations also
    require the configured frontend Origin. Health/authentication and preflight
    are exempt. Local mode is unauthenticated trusted development, not an
    alternate authentication mechanism. HTTPS uses __Host-agentflow_session;
    loopback HTTP uses agentflow_session. Archived business writes return 409;
    missing scope resolves to the owner's active default. Workspace management
    is owner-scoped independently of the selected header. Agent/Tool configuration
    writes remain trusted-local operator operations (OIDC users receive 403).
    Resource IDs are resolved inside that authorized Workspace; foreign IDs return
    404. SSE delivery rechecks the session and Workspace on each write, including
    Chat/Continue/Resume. A later access failure emits an error frame using the
    existing ChatStreamEvent error contract and stops delivery, without canceling
    the admitted durable Run. See the resource authorization guide.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/agents:
    get:
      operationId: listAgents
      tags: [Agents]
      responses:
        '200':
          description: Active agents.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createAgent
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentConfigRequest'
      responses:
        '201':
          description: Created agent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/agents/{id}:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    get:
      operationId: getAgent
      tags: [Agents]
      responses:
        '200':
          description: Agent configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateAgent
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentConfigRequest'
      responses:
        '200':
          description: Updated agent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: archiveAgent
      tags: [Agents]
      responses:
        '204':
          description: Agent archived.
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/runs:
    get:
      operationId: listRuns
      tags: [Runs]
      responses:
        '200':
          description: Runs ordered by the store.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Run'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/runs/{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/RunId'
    post:
      operationId: cancelRun
      tags: [Runs]
      responses:
        '200':
          description: Canceled or canceling run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/runs/{id}/replay:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      operationId: getRunReplay
      tags: [Runs]
      responses:
        '200':
          description: Replayable Run state and evidence.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunReplay'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/runs/{id}/usage:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      operationId: getRunUsage
      tags: [Runs]
      responses:
        '200':
          description: Run budget and append-only usage ledger.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunUsageLedger'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/runs/{id}/collaboration_steps:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      operationId: listCollaborationSteps
      tags: [Runs]
      responses:
        '200':
          description: Multi-Agent or Autonomous stages for the Run.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollaborationStep'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
...
    AgentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    RunId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current resource state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        source:
          type: string
        category:
          type: string
        retryable:
          type: boolean
        operation:
          type: string
        request_id:
          type: string
...
    RunStatus:
      type: string
      enum: [queued, running, waiting_for_user, completed, failed, failed_recoverable, canceling, canceled]
    VerificationStatus:
      type: string
      enum: [not_required, pending, running, passed, failed, blocked, stale]
...
    VerificationPolicyInput:
      type: object
      required: [mode, max_attempts, on_exhausted]
      properties:
        mode:
          type: string
          enum: [all_must_pass, any_may_pass]
        max_attempts:
          type: integer
          minimum: 1
        on_exhausted:
          type: string
          enum: [fail, waiting_for_user]
    CompletionContractInput:
      type: object
      required: [subject_type, verifiers, policy]
      properties:
        subject_type:
          type: string
          enum: [run_output]
        verifiers:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/VerifierSpecInput'
        policy:
          $ref: '#/components/schemas/VerificationPolicyInput'
    CompletionContract:
      allOf:
        - $ref: '#/components/schemas/CompletionContractInput'
        - type: object
          required: [id, version, hash]
          properties:
            id:
              type: string
            version:
              type: integer
            hash:
              type: string
...
    Conversation:
      type: object
      required: [id, workspace_id, title, created_at, updated_at]
      properties:
        id:
          type: string
        workspace_id:
          type: string
        title:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Message:
      type: object
      required: [id, workspace_id, conversation_id, role, content, created_at]
      properties:
        id:
          type: string
        workspace_id:
          type: string
        conversation_id:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        citations:
          type: array
          items:
            $ref: '#/components/schemas/RAGCitation'
        web_citations:
          type: array
          items:
            $ref: '#/components/schemas/WebCitation'
        reasoning:
          type: array
          description: Sanitized completed provider reasoning associated with this assistant answer; display-only, not model input.
          items:
            $ref: '#/components/schemas/MessageReasoning'
        created_at:
          type: string
          format: date-time
    RAGCitation:
      type: object
      required: [source_id, document_id, document_title, chunk_id]
      properties:
        source_id: { type: string }
        document_id: { type: string }
        document_title: { type: string }
        document_version: { type: string }
        chunk_id: { type: string }
        source_chunk_ids:
          type: array
          items: { type: string }
        section_path:
          type: array
          items: { type: string }
        start_offset: { type: integer }
        end_offset: { type: integer }
        run_id: { type: string }
        tool_call_id: { type: string }
        tool_event_id: { type: string }
    WebCitation:
      type: object
      required: [source_id, title, url, run_id, tool_call_id, tool_event_id]
      properties:
        source_id: { type: string }
        title: { type: string }
        url: { type: string, format: uri }
        run_id: { type: string }
        tool_call_id: { type: string }
        tool_event_id: { type: string }
    AgentConfigRequest:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        system_prompt:
          type: string
        routing_hints:
          $ref: '#/components/schemas/AgentRoutingHints'
        tools:
          type: array
          items:
            type: string
        skills:
          type: array
          maxItems: 8
          items: { type: string }
        memory_enabled:
          type: boolean
        retrieval_enabled:
          type: boolean
    Agent:
      type: object
      required: [id, name, description, system_prompt, tools, memory_enabled, retrieval_enabled, created_at, updated_at]
      properties:
        workspace_id:
          type: string
          description: Owning Workspace; absent for shared read-only templates.
        is_template:
          type: boolean
          description: Shared read-only profile; POST a copy to configure it in a Workspace.
        id:
          type: string
        name:
          type: string
        description:
          type: string
        system_prompt:
          type: string
        routing_hints:
          $ref: '#/components/schemas/AgentRoutingHints'
        tools:
          type: array
          items:
            type: string
        skills:
          type: array
          items: { type: string }
        memory_enabled:
          type: boolean
        retrieval_enabled:
          type: boolean
        executor:
          type: string
        archived:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AgentRoutingHints:
      type: object
      required: [capabilities, task_examples, exclusions]
      properties:
        capabilities:
          type: array
          items:
            type: string
        task_examples:
          type: array
          items:
            type: string
        exclusions:
          type: array
          items:
            type: string
...
    Run:
      type: object
      required: [id, workspace_id, agent_id, conversation_id, status, verification_status, active_runtime_ms, created_at, updated_at]
      properties:
        id:
          type: string
        workspace_id:
          type: string
        agent_id:
          type: string
        conversation_id:
          type: string
        status:
          $ref: '#/components/schemas/RunStatus'
        verification_status:
          $ref: '#/components/schemas/VerificationStatus'
        completion_contract:
          $ref: '#/components/schemas/CompletionContract'
        error:
          type: string
        started_at:
          type: string
          format: date-time
        execution_started_at:
          type: string
          format: date-time
        active_runtime_ms:
          type: integer
          format: int64
        heartbeat_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CollaborationStep:
      type: object
      required: [id, run_id, conversation_id, role, status, input, output, created_at, updated_at]
      properties:
        id:
          type: string
        run_id:
          type: string
        conversation_id:
          type: string
        role:
          type: string
        agent_id:
          type: string
        status:
          type: string
        iteration:
          type: integer
        input:
          type: string
        output:
          type: string
        error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RunEvent:
      type: object
      required: [id, type, schema_version, sequence, run_id, payload, timestamp]
      properties:
        id:
          type: string
        type:
          type: string
        schema_version:
          type: integer
        sequence:
          type: integer
          format: int64
        conversation_id:
          type: string
        run_id:
          type: string
        stage_id:
          type: string
        turn_id:
          type: string
        parent_event_id:
          type: string
        payload:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
    RunTraceSummary:
      type: object
      required: [run_id, status, total_duration_ms, total_tokens, prompt_tokens, completion_tokens, token_usage_estimated, llm_calls, tool_calls, error_count]
      properties:
        run_id:
          type: string
        status:
          $ref: '#/components/schemas/RunStatus'
        total_duration_ms:
          type: integer
          format: int64
        total_tokens:
          type: integer
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        token_usage_estimated:
          type: boolean
        llm_calls:
          type: integer
        tool_calls:
          type: integer
        error_count:
          type: integer
    RuntimeRunBudget:
      type: object
      properties:
        max_model_calls: { type: integer }
        max_prompt_tokens: { type: integer }
        max_completion_tokens: { type: integer }
        max_total_tokens: { type: integer }
        max_tool_calls: { type: integer }
        max_runtime_ms: { type: integer, format: int64 }
        max_estimated_cost_micros: { type: integer, format: int64 }
        input_cost_per_million_tokens_micros: { type: integer, format: int64 }
        output_cost_per_million_tokens_micros: { type: integer, format: int64 }
    RunUsageTotals:
      type: object
      required: [model_calls, tool_calls, prompt_tokens, completion_tokens, total_tokens, estimated_cost_micros, open_reservations]
      properties:
        model_calls: { type: integer }
        tool_calls: { type: integer }
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        total_tokens: { type: integer }
        estimated_cost_micros: { type: integer, format: int64 }
        open_reservations: { type: integer }
        cost_unknown_entries: { type: integer, minimum: 0 }
...
    RunUsageEntry:
      type: object
      required: [id, run_id, operation_id, kind, purpose, timestamp]
      properties:
        id: { type: string }
        run_id: { type: string }
        operation_id: { type: string }
        stage_id: { type: string }
        turn_id: { type: string }
        kind:
          type: string
          enum: [model.reservation, model.settlement, tool.execution]
        purpose:
          type: string
          enum: [primary, router, compaction]
        model: { type: string }
        tool_name: { type: string }
        model_calls: { type: integer }
        tool_calls: { type: integer }
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        total_tokens: { type: integer }
        estimated_cost_micros: { type: integer, format: int64 }
        estimated: { type: boolean }
        breakdown: { $ref: '#/components/schemas/UsageBreakdown' }
        cost_details: { $ref: '#/components/schemas/UsageCostDetails' }
        timestamp: { type: string, format: date-time }
    RunUsageLedger:
      type: object
      required: [run_id, budget, totals, entries]
      properties:
        run_id: { type: string }
        budget:
          $ref: '#/components/schemas/RuntimeRunBudget'
        totals:
          $ref: '#/components/schemas/RunUsageTotals'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/RunUsageEntry'
        updated_at:
          type: string
          format: date-time
    VerificationEvidence:
      type: object
      additionalProperties: true
      required: [id, run_id, verifier_id, status]
      properties:
        id: { type: string }
        run_id: { type: string }
        verifier_id: { type: string }
        status:
          $ref: '#/components/schemas/VerificationStatus'
    VerificationArtifact:
      type: object
      additionalProperties: true
      required: [id, run_id, evidence_id, kind]
      properties:
        id: { type: string }
        run_id: { type: string }
        evidence_id: { type: string }
        kind: { type: string }
    SkillResourceEvidence:
      type: object
      required: [path, hash, offset, next_offset, total_bytes, event_id, sequence]
      properties:
        path: { type: string }
        hash: { type: string }
        offset: { type: integer }
        next_offset: { type: integer }
        total_bytes: { type: integer }
        event_id: { type: string }
        sequence: { type: integer, format: int64 }
    SkillFailureEvidence:
      type: object
      required: [tool, code, event_id, sequence]
      properties:
        tool: { type: string }
        code: { type: string }
        path: { type: string }
        event_id: { type: string }
        sequence: { type: integer, format: int64 }
    SkillEvidence:
      type: object
      required: [name, hash, agent_id, bound, instructions, activation, resources, failures]
      properties:
        name: { type: string }
        hash: { type: string }
        agent_id: { type: string }
        stage_id: { type: string }
        bound: { type: boolean }
        instructions: { type: string, enum: [included, not_observed] }
        activation: { type: string, enum: [explicit, model, not_observed] }
        first_sequence: { type: integer, format: int64 }
        first_request_sequence: { type: integer, format: int64 }
        manifest_id: { type: string }
        event_id: { type: string }
        request_id: { type: string }
        estimated_tokens: { type: integer }
        resources:
          type: array
          items: { $ref: '#/components/schemas/SkillResourceEvidence' }
        failures:
          type: array
          items: { $ref: '#/components/schemas/SkillFailureEvidence' }
    PartialOutput:
      description: Bounded display replacement committed at an event sequence; never model continuation or a final Message.
      type: object
      required: [run_id, turn_id, channel, round, revision, offset, text, status, sequence]
      properties:
        run_id: { type: string }
        stage_id: { type: string }
        turn_id: { type: string }
        model_call_id: { type: string }
        attempt: { type: integer, minimum: 0 }
        round: { type: integer, minimum: 1 }
        channel: { type: string, enum: [answer, reasoning] }
        role: { type: string }
        revision: { type: integer, format: int64, minimum: 1 }
        offset: { type: integer, minimum: 0, description: UTF-8 byte length of this display replacement, not provider stream offset. }
        text: { type: string }
        status: { type: string, enum: [provisional, retracted, final, interrupted] }
        truncated: { type: boolean }
        sequence: { type: integer, format: int64, minimum: 0 }
    ToolProgressUpdate:
      description: Bounded semantic progress only; counts do not imply Tool completion.
      type: object
      required: [phase]
      properties:
        phase: { type: string, pattern: '^[a-z0-9_-]{1,64}$' }
        message: { type: string, maxLength: 1024, description: Backend enforces a 1024 UTF-8 byte cap after redaction. }
        completed: { type: integer, format: int64, minimum: 0 }
        total: { type: integer, format: int64, minimum: 1 }
        truncated: { type: boolean }
    ToolProgress:
      allOf:
        - $ref: '#/components/schemas/ToolProgressUpdate'
        - type: object
          required: [run_id, turn_id, tool_call_id, tool_name, status, sequence]
          properties:
            run_id: { type: string }
            turn_id: { type: string }
            stage_id: { type: string }
            tool_call_id: { type: string }
            tool_name: { type: string }
            status: { type: string, enum: [running, completed, failed, interrupted] }
            sequence: { type: integer, format: int64 }
...
    RunReplay:
      type: object
      required: [run, projection, conversation, messages, steps, summary, usage_ledger, run_events, stage_checkpoints, tool_effects, tool_artifacts, verification_evidence, verification_artifacts, task_state_revisions]
      properties:
        run:
          $ref: '#/components/schemas/Run'
        projection:
          type: object
          additionalProperties: true
          properties:
            skill_evidence:
              type: array
              items: { $ref: '#/components/schemas/SkillEvidence' }
            partial_outputs:
              type: array
              items: { $ref: '#/components/schemas/PartialOutput' }
            tool_progress:
              type: array
              items: { $ref: '#/components/schemas/ToolProgress' }
        runtime_snapshot:
          type: object
          additionalProperties: true
        conversation:
          $ref: '#/components/schemas/Conversation'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/CollaborationStep'
        summary:
          $ref: '#/components/schemas/RunTraceSummary'
        usage_ledger:
          $ref: '#/components/schemas/RunUsageLedger'
        run_events:
          type: array
          items:
            $ref: '#/components/schemas/RunEvent'
        stage_checkpoints:
          type: array
          items:
            type: object
            additionalProperties: true
        tool_effects:
          type: array
          items:
            type: object
            additionalProperties: true
        tool_artifacts:
          type: array
          items:
            type: object
            additionalProperties: true
        verification_evidence:
          type: array
          items:
            $ref: '#/components/schemas/VerificationEvidence'
        verification_artifacts:
          type: array
          items:
            $ref: '#/components/schemas/VerificationArtifact'
        task_state_revisions:
          type: array
          items:
            type: object
            additionalProperties: true
        recovery_summary:
          type: object
          additionalProperties: true
...